[nasa/cryptolib#145] Added initial version of Key Interface;
