+++ CMakeLists.txt
@@ -27,6 +27,9 @@ option(CODECOV "Code Coverage" OFF)
 option(CRYPTO_LIBGCRYPT "Cryptography Module - Libgcrypt" ON)
 option(CRYPTO_KMC "Cryptography Module - KMC" OFF)
 option(DEBUG "Debug" OFF)
+option(KEY_CUSTOM "Key Module - Custom" OFF)
+option(KEY_INTERNAL "Key Module - Internal" ON)
+option(KEY_KMC "Key Module - KMC OFF")
 option(SA_INTERNAL "Security Association - Internal" ON)
 option(SA_MARIADB "Security Association - MariaDB" OFF)
 option(SUPPORT "Support" OFF)
