+++ src/CMakeLists.txt
@@ -35,6 +35,31 @@ else()
     list(APPEND LIB_SRC_FILES ${KMC_FILES})
 endif()
 
+if(KEY_CUSTOM)
+    # Assumes CryptoLib is a Git submodule to project and custom directories and definitions exist at top level
+    aux_source_directory(../../key/custom KEY_CUSTOM_FILES)
+    list(APPEND LIB_SRC_FILES ${KEY_CUSTOM_FILES})
+else()
+    aux_source_directory(key/custom_stub KEY_CUSTOM_FILES)
+    list(APPEND LIB_SRC_FILES ${KEY_CUSTOM_FILES})
+endif()
+
+if(KEY_INTERNAL)
+    aux_source_directory(key/internal KEY_INTERNAL_FILES)
+    list(APPEND LIB_SRC_FILES ${KEY_INTERNAL_FILES})
+else()
+    aux_source_directory(key/internal_stub KEY_INTERNAL_FILES)
+    list(APPEND LIB_SRC_FILES ${KEY_INTERNAL_FILES})
+endif()
+
+if(KEY_KMC)
+    aux_source_directory(key/kmc KEY_KMC_FILES)
+    list(APPEND LIB_SRC_FILES ${KEY_KMC_FILES})
+else()
+    aux_source_directory(key/kmc_stub KEY_KMC_FILES)
+    list(APPEND LIB_SRC_FILES ${KEY_KMC_FILES})
+endif()
+
 if(SA_INTERNAL)
     aux_source_directory(sa/internal SA_INTERNAL_FILES)
     list(APPEND LIB_SRC_FILES ${SA_INTERNAL_FILES})
