+++ src/key/internal_stub/key_interface_internal_stub.template.c
@@ -1,0 +1,29 @@
+/* Copyright (C) 2009 - 2022 National Aeronautics and Space Administration.
+   All Foreign Rights are Reserved to the U.S. Government.
+
+   This software is provided "as is" without any warranty of any kind, either expressed, implied, or statutory,
+   including, but not limited to, any warranty that the software will conform to specifications, any implied warranties
+   of merchantability, fitness for a particular purpose, and freedom from infringement, and any warranty that the
+   documentation will conform to the program, or any warranty that the software will be error free.
+
+   In no event shall NASA be liable for any damages, including, but not limited to direct, indirect, special or
+   consequential damages, arising out of, resulting from, or in any way connected with the software or its
+   documentation, whether or not based upon warranty, contract, tort or otherwise, and whether or not loss was sustained
+   from, or arose out of the results of, or use of, the software, documentation or services provided hereunder.
+
+   ITC Team
+   NASA IV&V
+   jstar-development-team@mail.nasa.gov
+*/
+
+#include "key_interface.h"
+
+/* Variables */
+static KeyInterfaceStruct key_if_struct;
+
+/* Functions */
+KeyInterface get_key_interface_internal(void)
+{
+    fprintf(stderr,"ERROR: Loading internal key interface stub source code. Rebuild CryptoLib with -DKEY_INTERNAL=ON to use implementation.\n");
+    return &key_if_struct;
+}
